modified:   plnode/message/control/peer_change_status_message.h 	modified:   plnode/message/control/peer_dyn_change_status_message.h 	modified:   plnode/message/control/peer_gen_name_message.h 	modified:   plnode/message/control/peer_init_message.h 	modified:   plnode/message/control/peer_initiate_get.h 	modified:   plnode/message/control/peer_initiate_put.h 	modified:   plnode/message/control/peer_start_message.h 	modified:   plnode/message/message.h 	modified:   plnode/message/p2p/message_put.h 	modified:   plnode/protocol/plexus/plexus_protocol.h 	modified:   server_system_test.cc

+++ plnode/message/control/peer_dyn_change_status_message.h
@@ -22,9 +22,9 @@ public:
 	{
 	}
 
-	int getSize()
+	size_t getSize()
 	{
-		int ret = getBaseSize();
+		size_t ret = getBaseSize();
 		return ret;
 	}
