+++ plnode/message/control/peer_start_message.h
@@ -20,9 +20,9 @@ public:
 	{
 	}
 
-	int getSize()
+	size_t getSize()
 	{
-		int ret = getBaseSize();
+		size_t ret = getBaseSize();
 		return ret;
 	}
